requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncclick==8.1.3.4
asyncpg==0.27.0
bcrypt==4.0.1
black==23.3.0
boto3==1.26.165
botocore==1.29.165
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.8.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.9
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
